==> R/00_all.qmd <==
---
title: "00_all"
format:
  html:
    embed-resources: true
editor: visual
---

```{r}
#| include: false
library(quarto)

# Render all analysis quarto documents
quarto_render(input = "01_load.qmd")
quarto_render(input = "02_clean.qmd")
quarto_render(input = "03_augment.qmd")
quarto_render(input = "04_describe.qmd")
quarto_render(input = "05_size_analysis.qmd")
quarto_render(input = "06_logoplot.qmd")
quarto_render(input = "07_location_analysis.qmd")

# Move all rendered files to results folder
file.rename(from = "01_load.html",
             to = "../results/01_load.html")
file.rename(from = "02_clean.html",
            to = "../results/02_clean.html")
file.rename(from = "03_augment.html",
            to = "../results/03_augment.html")
file.rename(from = "04_describe.html",
            to = "../results/04_describe.html")
file.rename(from = "05_size_analysis.html",
            to = "../results/05_size_analysis.html")
file.rename(from = "06_logoplot.html",
            to = "../results/06_logoplot.html")
file.rename(from = "07_location_analysis.html",
            to = "../results/07_location_analysis.html")
```

{{< include 01_load.qmd >}}

{{< include 02_clean.qmd >}}

{{< include 03_augment.qmd >}}

{{< include 04_describe.qmd >}}

{{< include 05_size_analysis.qmd >}}

{{< include 06_logoplot.qmd >}}

{{< include 07_location_analysis.qmd >}}
