==> R/03_augment.qmd <==
---
title: ""
format:
  html:
    embed-resources: true
editor: visual
---

# Augmentation of data

## Load librares

```{r}
#| message: false
library(tidyverse)
```

## Load clean data

```{r}
#| message: false
clean_data <- read_delim(file = "../data/02_dat_clean.tsv", delim = "\t")
```

## Augment data

### Add column with peptide sequence around active site

```{r}
# Function add column with sequence around active site
get_logo_seq <- function(x, y) {
  y <- as.integer(y)
  logo_seq_length <- 7
  n_aa_per_side <- (logo_seq_length-1/2)
  return(substr(x, y-3, y+3))
}

# Extracting peptides around active site and saving in new column
aug_data <- clean_data |>
  mutate(logo_seq = get_logo_seq(sequence, active_site)) |>
  # Filtereing out k-mers not covering 7 AA's
  mutate(logo_seq = if_else(nchar(logo_seq) != 7, NA, logo_seq))
```

### Add count for the number of molecular functions for each protein

```{r}
####
# Count the number of molecular functions for each entry
####
molecular_function_count <- clean_data |>
  select(organism, entry, GO_molecular_function, length, mass) |>
  distinct() |>
  drop_na() |>
  count(entry, name = "num_molecular_functions") 

# Join num_molecular_functions
aug_data <- aug_data |>
  full_join(molecular_function_count, by = "entry")

```

### Add number of cofactors that binds to each protein

```{r}
####
# Count the number of cofactors
####
cofactor_count <- clean_data |>
  select(organism, entry, cofactor, length, mass) |>
  distinct() |>
  drop_na() |>
  count(entry, name = "num_cofactors") 

# Join num_molecular_functions
aug_data <- aug_data |>
  full_join(cofactor_count, by = "entry")

```

### Convert the first EC number into an enzyme class number

```{r}
####
# Change "EC" column (we have decided only to work with the first digit but we are aware that the others have information aswell)
####
aug_data <- aug_data |>
  # Extract the first digit and create new column
  # For the few cases with more ec numbers we just look at the first
  mutate(first_digit = str_extract(ec_number, "^\\d"),
         # Translate digit to enzyme class using vector
         enzyme_class = case_when(
           first_digit == 1  ~ "oxidoreductases",
           first_digit == 2  ~ "transferases",
           first_digit == 3  ~ "hydrolases",
           first_digit == 4  ~ "lyases",
           first_digit == 5  ~ "isomerases",
           first_digit == 6  ~ "ligases",
           first_digit == 7  ~ "translocases",
         )) |> 
  select(-first_digit, -ec_number)

```

## Save augment data object

```{r}
write_tsv(aug_data, file = "../data/03_dat_aug.tsv")
```
