==> R/05_size_analysis.qmd <==
---
title: ""
format:
  html:
    embed-resources: true
editor: visual
---

# Protein size analysis

## Load augmented data

```{r}
#| message: false
library(tidyverse)
# Load data
data <- read_delim(file = "../data/03_dat_aug.tsv", delim = "\t")
```

## Plot number of molecular functions vs length

We wanted to see if proteins with more molecular functions were larger

```{r}
#####
# Plot counted number of molecular functions vs length 
#####
data |>
  select(organism,
         entry,
         length,
         num_molecular_functions) |>
  distinct() |>
  drop_na() |>
  ggplot(aes(x = length,
             y = num_molecular_functions,
             fill = organism,
             colour = organism)) + 
  geom_point(alpha = 0.5,
             size = 0.75)+
  labs(title = "Length vs. Molecular Function Count",
       x = "Length",
       y = "Number of Molecular Functions") +
  theme_bw()

ggsave("05_size_analysis_mol_func.png",
       path = "../results")
```

## Boxplot of cofactor vs length

We wanted to see if proteins that had more cofactors were generally larger. We chose to do this as a boxplot because the maximum number of cofactors was 5 and it was therefore a visually manageable plot that eased visualization substantially. it did however require that the number of cofactors was treated as a factor.

```{r}
#####
# Plot counted number of cofactors functions vs length 
#####
data |>
    select(organism,
           entry,
           length,
           num_cofactors)|>
    distinct()|>
    drop_na()|>
    # num_cofactors is converted to factor to enable use of boxplot
    ggplot(aes(y = length,
               fill = factor(num_cofactors),
               colour = factor(num_cofactors))) + 
    geom_boxplot(alpha = 0.5,
                 size = 0.75)+
    labs(title = "Length by number of binding cofactors",
         y = "Length",
         fill = "Number of cofactors",
         color = "Number of cofactors") +
    facet_wrap(~organism)+
    # Extreme outliers are ignored to improve visualization
    scale_y_continuous(limits = c(0, quantile(data$length, 0.98))) +
    theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  theme_bw()

ggsave("05_size_analysis_cofactor_size.png",
       path = "../results")
```

## Plot size vs mass in kDa

We were wondering if humans used more large side chains than mice. To investigate this we plotted length vs mass and performed linear regression. We made one linear model for each organism and plotted the 95% confidence interval as to enable easy visualization to check if they were significantly different from each other. Should this be the case it would indicate that the two species do in fact have differences in the size of their amino acid side chains.

```{r}
#####
# Plot size vs mass
#####
# Linear model of mass vs length for mice
mouse_model <- lm(mass ~ length,
   data = data |>
     select(organism,
            entry,
            length,
            mass) |>
     filter(organism == "Mouse") |>
     distinct() |>
     drop_na())


summary(mouse_model)

# Linear model for mass vs length for humans
human_model <- lm(mass ~ length,
   data = data |>
     select(organism,
            entry,
            length,
            mass) |>
     filter(organism == "Human") |>
     distinct() |>
     drop_na())



summary(human_model)

# combined linear model
combined_linear_model <- lm(mass ~ length * organism,
                   data = data |>
                             select(organism,
                                    entry,
                                    length,
                                    mass) |>
                             distinct() |>
                             drop_na())
                    
summary <- summary(combined_linear_model)

length_mouse <-summary$coefficients[4,1]
pval_mouse <- summary$coefficients[4,4]

data |>
  select(organism,
       entry,
       length,
       mass) |>
  distinct() |>
  drop_na() |>
  group_by(organism) |>
  # Create scatter plot with linear regression lines
  ggplot(aes(x = length,
             y = mass,
             color = organism,
             linetype = organism)) +
    geom_point(alpha = 0.25,
               size = 0.75) +
    # Add linear regression line for human
    geom_smooth(method = "lm",
                se = TRUE,
                formula = y ~ x) +
    geom_text(x = 3000, 
              y = 100000, 
              label = paste("Difference in slope =", length_mouse),
              size = 3.5, 
              color = "black")+
    geom_text(x = 3000, 
                y = 50000, 
                label = paste("p value =", pval_mouse),
                size = 3.5, 
                color = "black")+
    labs(title = "Length vs. Mass",
        subtitle = "Length vs. mass of proteins in humans and mice including linear regressions",
        x = "Length (amino acids)",
        y = "Mass (kDa)",
        color = "Organism",
        linetype = "Organism") +
    # Extreme outliers are ignored to improove visualization
    scale_y_continuous(labels = function(x) format(x, scientific = FALSE),
                       limits = c(0, quantile(data$mass, 0.99))) +
    scale_x_continuous(labels = function(x) format(x, scientific = FALSE),
                       limits = c(0, quantile(data$length, 0.99)))+
    theme_bw()

ggsave("05_size_analysis_lm_size_length.png",
      path = "../results")
```

The linear models for humans and mice are very similar with the slope of humans being 110.50162 kDa while mice have 110.78923. When running the combined linear model, we can however see that this small change in the slope is statistically significant with a p-value of 4.39\*e-10. This highly significant difference of 0.28760 kDa in the slope indicates that mice do in fact have slightly larger side chains than humans. The difference in slope is harder to draw meaningful conclusions on as this is the supposed mass when the length is zero. The fact that the intercept is not zero is most likely to improve model fit.

## pH Length investigation

The idea was to see if there was any connection between protein size and the optimal pH for the protein. The datapoints are coloured by organism to see if there are any clear differences between mice and humans.

```{r}
####
# Assemble size-pH
####
data|>
  select(organism,
         entry,
         length,
         mass,
         ph_dependence)|>
  distinct()|>
  drop_na()|>
  # 368 distinct entries
  # Plot a scatter plot of the 
  ggplot(aes(x= length,
             y = ph_dependence,
             fill = organism,
             colour = organism)) + 
  geom_point(alpha = 0.75,
             size = 0.75)+
  labs(title = "Scatter Plot of length vs. pH",
       x = "Sequence length",
       y = "pH optimum") +
  theme_bw()

ggsave("05_size_analysis_pH_length.png",
       path = "../results")
```

## pH enzyme class comparison

We decided to test if enzyme classes had varying pH dependencies between humans and mice. We therefore made a boxplot of the pH dependencies on all enzyme classes and stratified on the organism to compare the two.

```{r}
####
# pH by enzyme class test 
####
data|>
  # Filter remoove unwanted columns and rows from data
  select(organism,
         entry,
         length,
         mass,
         ph_dependence,
         enzyme_class) |>
  distinct() |>
  drop_na() |>
  # Plot the results  
  ggplot(aes(y = ph_dependence, 
             fill = enzyme_class, 
             colour = enzyme_class)) +
  geom_boxplot(alpha = 0.75, size = 0.9) +
  labs(color = "Enzyme class",
       fill = "Enzyme class",
       y = "pH optimum") +
  facet_wrap(~organism) +
  theme_bw()
   
# 310 distinct entries
data|>
  select(organism,
         entry,
         ph_dependence,
         enzyme_class)|>
  filter(organism == "Mouse",
         enzyme_class == "ligases")|>
  drop_na()|>
  distinct(entry)
# Only 5 ligases in mice which is not enough to determine anything 
```

## Plotting Length vs enzyme class

The idea was to see if mice and human proteins varied in size when looking at specific enzyme classes.

```{r}
####
# Size by enzyme class test 
####
data|>
    select(organism,
           entry,
           length,
           mass,
           enzyme_class) |>
    distinct() |>
    drop_na() |>
    ggplot(aes(y = length,
               fill = enzyme_class,
               colour = enzyme_class)) +
    geom_boxplot(alpha = 0.75, size = 0.9) +
    labs(title = "Length by enzyme class",
         color = "Enzyme class",
         fill = "Enzyme class",
         y = "Length") + 
    facet_wrap(~organism) +
    scale_y_continuous(limits = c(0, quantile(data$length, 0.98))) +
    theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  theme_bw()
# 310 distinct entries

ggsave("05_size_analysis_class_size.png",
       path = "../results")
```

## Plotting size by subcellular loaction

The idea was to see if protein sizes varied in subcellular locations and to check if there were any differences between mice and humans.

```{r}
####
# Size by subcellular location 
####

# Plot boxplots for protein sizes in three subcellular locations
data |>
  filter(`subcellular_location_[cc]` == c("cell membrane", "cytoplasm","nucleus")) |>
  ggplot(aes(y = length,
             fill = `subcellular_location_[cc]`,
             colour = `subcellular_location_[cc]`)) +
  geom_boxplot(alpha = 0.75, size = 0.9) +
  facet_wrap(~organism) +
  scale_y_continuous(limits = c(0, quantile(data$length, 0.99))) +
  labs(title = "Boxplot of protein lengths in three subcellular locations",
       y = "Protein Length",
       fill = "Subcellular location",
       color = "Subcellular location") +
  theme(legend.position = "bottom") +
  theme_bw()

ggsave("05_size_analysis_location_size.png",
       path = "../results")
```
